bundle functions into preprocess.py
